Rename URI2 to URI
