+++ lib/json-schema/schema.rb
@@ -10,13 +10,13 @@ module JSON
 
       # If there is an ID on this schema, use it to generate the URI
       if @schema['id'].is_a?(String)
-        temp_uri = JSON::Util::URI2.parse(@schema['id'])
+        temp_uri = JSON::Util::URI.parse(@schema['id'])
         if temp_uri.relative?
           temp_uri = uri.join(temp_uri)
         end
         @uri = temp_uri
       end
-      @uri = JSON::Util::URI2.strip_fragment(@uri)
+      @uri = JSON::Util::URI.strip_fragment(@uri)
 
       # If there is a $schema on this schema, use it to determine which validator to use
       if @schema['$schema']
