+++ lib/json-schema/validators/draft6.rb
@@ -44,7 +44,7 @@ module JSON
           'uri' => UriFormat,
         }
         @formats = @default_formats.clone
-        @uri = JSON::Util::URI2.parse('http://json-schema.org/draft-06/schema#')
+        @uri = JSON::Util::URI.parse('http://json-schema.org/draft-06/schema#')
         @names = ['draft6', 'http://json-schema.org/draft-06/schema#']
         @metaschema_name = 'draft-06.json'
       end
