+++ lib/json-schema/validators/hyper-draft2.rb
@@ -3,7 +3,7 @@ module JSON
     class HyperDraft2 < Draft2
       def initialize
         super
-        @uri = JSON::Util::URI2.parse('http://json-schema.org/draft-02/hyper-schema#')
+        @uri = JSON::Util::URI.parse('http://json-schema.org/draft-02/hyper-schema#')
       end
 
       JSON::Validator.register_validator(new)
