+++ lib/json-schema/validators/hyper-draft3.rb
@@ -3,7 +3,7 @@ module JSON
     class HyperDraft3 < Draft3
       def initialize
         super
-        @uri = JSON::Util::URI2.parse('http://json-schema.org/draft-03/hyper-schema#')
+        @uri = JSON::Util::URI.parse('http://json-schema.org/draft-03/hyper-schema#')
       end
 
       JSON::Validator.register_validator(new)
