+++ lib/json-schema/validators/hyper-draft4.rb
@@ -3,7 +3,7 @@ module JSON
     class HyperDraft4 < Draft4
       def initialize
         super
-        @uri = JSON::Util::URI2.parse('http://json-schema.org/draft-04/hyper-schema#')
+        @uri = JSON::Util::URI.parse('http://json-schema.org/draft-04/hyper-schema#')
       end
 
       JSON::Validator.register_validator(new)
