+++ lib/json-schema/validators/hyper-draft6.rb
@@ -3,7 +3,7 @@ module JSON
     class HyperDraft6 < Draft6
       def initialize
         super
-        @uri = JSON::Util::URI2.parse('http://json-schema.org/draft-06/hyper-schema#')
+        @uri = JSON::Util::URI.parse('http://json-schema.org/draft-06/hyper-schema#')
       end
 
       JSON::Validator.register_validator(new)
